added a couple toy BDA scripts
- modelcomparison_1.js:

in the "binary choice", "animals A + B on card", "speaker says 'AorB" condition, we had 114 observations, 100 of which were "Right". Given this, what is the likelihood of our literal listener model, compared against an alternative model (in this case, the speaker model)?

In this case, P is close to 1: the speaker model does a really bad job reative to the null model of predicting the data we saw.

- parameterestimation_1.js:

estimate our rationality parameter given our speaker model and the observed data (same as above).